MONGOCRYPT-210 fix static builds on windows
Since kms_message is built as a static library, it needs to be
included on the link line for all statically compiled targets that
need to include it. Additionally, `KMS_MSG_STATIC` needed to be
defined during `mongocrypt_static` compilation in order to avoid
prefixing symbols which assumed a dynamic build.