feat: support Azure and GCP KMS providers
This patch set updates the Node.js bindings for libmongocrypt to
support Azure and GCP as KMS providers for CSFLE. The changes
should be straightforward, following the migration guide for
this version of libmongocrypt. Notably, the KMS client code
introduced a dependency on `BufferList` in order to simplify the
logic of reading in chunks, and forwarding those chunks to
libmongocrypt.

NODE-2825