MONGOCRYPT-334 clone mongo-c-driver from reference in all tests
libmongocrypt pins to 1.17.0 of the C driver when building in
prep_c_driver_source.sh.

linker-tests.sh and pkgconfig-tests.sh work off of the tip of
master for the C driver.

libmongocrypt tests fail to compile on RHEL 6 / gcc 4.4 due to
a dependency on newer atomic builtins in CDRIVER-4124.

This updates tests to build off using C 1.17.0.