fix(NODE-4279): handle decrypted nested documents well with devtools decoration
Handle cases in which the decrypted values are documents/arrays
in the devtools-specific decrypted-reponse marking.

This was previously untested because it’s quite a bit of effort
to generate new mock data (like the one added in this commit)
for the current test setup, but obviously that was a mistake.