Suppress leak warning in OpenSSL with one shared file (#402)
This fixes the failing test execution on SLES 15. The suppression filepath option was set to a relative path. This meant that the suppression used were sensitive to the directory in which an executable runs.

This change sets the suppression filepath to point to an absolute path in the root of the repository.