Invoke CTest directly as part of the build script.
This removes the separate test.sh step and instead defines CTest tests
for each of the programs that we wish to execute for testing. We then
run CTest after each of the main build types we create. This will
increase configuration coverage, simplify the CI process, and expedite
test failures in the CI process.

The GCP and Azure tests are now executed as well, but emit test-skip
markers if the required environment variables are not defined.