MONGOCRYPT-681 sign Windows upload (#803)
* remove `windows-upload-check` task

Combines `windows-upload-check` and `windows-upload` into `windows-upload-release`.

Intended to ease testing. `windows-upload` used to only run for release commits. It now always runs. The version is appended to non-release uploads to separate.

* sign and upload

* deprecate fixed URLs

The URLs are overwritten on each release. Overwriting may add risk that an incorrect version is uploaded or downloaded.

* run `windows-upload-release` on Ubuntu 22.04

Task requires `docker` on the host. Windows does not appear to have `docker` installed. The `windows-upload-release` task does not require a Windows host.

* use Earthly

* use `.asc` file extension (not `.sig`) and verify signature when signing

* fix comment typo

Co-authored-by: Roberto C. Sánchez <roberto@connexer.com>

---------

Co-authored-by: Roberto C. Sánchez <roberto@connexer.com>