MONGOCRYPT-685 update libbson to 1.27.1 (#818)
* update bundled libbson to 1.27.1

* override `BUILD_VERSION` for C driver

If a `BUILD_VERSION` is specified for libmongocrypt, do not apply the same value for the embedded C driver.

* Clarify `ENABLE_ONLINE_TESTS` requires libmongoc. libmongoc is not required to build libmongocrypt.

* use the now-available `BSON_CONCAT` and remove TODOs

* remove `BSON_SUBTYPE_ENCRYPTED` (now defined in libbson)

* remove workaround for CDRIVER-3340

The removed code adds a runtime requirement of libbson 1.16.0 or newer. Add a check to ensure libbson 1.16.0 or newer is used at compile-time. This may prevent use of an older system install of libbson.

* disable building tests and counters in C driver

* update version in `prep_c_driver_source.sh`

* remove unnecessary VERSION_CURRENT

VERSION_CURRENT is now committed into the C driver

* add a `make_patch_command` CMake helper use for libbson and IntelDFP

* patch libbson to remove GCC diagnostic pragma

To resolve build error on GCC 4.4.7 on RHEL 6.2

* define `__STDC_LIMIT_MACROS`

To resolve build error on RHEL 6.2

* apply patch in linker tests

* reference ticket to drop RHEL 6.2

---------

Co-authored-by: vector-of-bool <vectorofbool@gmail.com>