MONGOCRYPT-661 enable "range" by default (#861)
* enable range v2 by default

unconditionally set `use_range_v2` to true

* update public header

Document "rangePreview" and "RangePreview" as removed. Using either results in an error.

Keep documentation of "range" as unstable. Protocol changes are in progress. "range" is expected to be stable for the upcoming 1.11.0 release.

* update test in Python bindings

expect error when "rangePreview" is used

* update Python C header include

* update test in Node bindings

expect an error when "rangePreview" is used

* update tests in Java bindings

* check for "Range" in Java binding validation

This check may be unnecessary. libmongocrypt validates options. However, the check is kept and updated to use "Range" to avoid changing the thrown exception.

* test "rangePreview" throws in Java bindings

* update Java bindings documentation

* fix Python format

* remove `enable_range_v2` option from `MongoCryptOptions` in Python bindings