DEVPROD-6063: update to macos 1100
This commit bumps our builds from MacOS 10.14 to MacOS 11. 10.14 has been
unsupported for several years. MacOS 11 is also EOL, but it seems to pass
all of our tests and isn't quite as old as 10.14. I tried to bump all the
way to MacOS 14, but ran into some kind of linking error. I think we can
return to that upgrade at a different time.