MONGOCRYPT-731 re-enable ASan tasks (#926)
* relocate misplaced comment.

Remove the comment from the YAML folded line to fix applying environment variables.

* suppress leak in dlerror.

To ignore a "Indirect leak" reported on Ubuntu 22.04 arm64. Assumed not to be an issue in libmongocrypt.

* remove unused `build-and-test-asan-s390x`

* document purpose of `build-and-test-asan-mac`

* set `ENABLE_EXTRA_ALIGNMENT=OFF` for imported libbson.

Fixes observed errors on older ASan with older gcc.

* print around tests in `pkgconfig-tests.sh` and `linker-tests.sh`.

To help identify a failing test.

* fix leaks in `_test_mc_FLE2TagAndEncryptedMetadataBlock_validate`

* document fix of disabled extra alignment. This may impact downstream consumers. Example: build with imported libbson, then load a different libbson at runtime.

This was the case in the pkgconfig tests. Disabling alignment was the default in 1.11 and accidentally changed in 1.12.0.

* apply `ENABLE_EXTRA_ALIGNMENT=OFF` in common CMake args for tests.

The `macos-1100` distro appears to have a system install of libbson with extra alignment enabled. libmongocrypt detects the system install and defaults to enable extra alignment to try to agree. This commit sets  `ENABLE_EXTRA_ALIGNMENT=OFF` for both libbson and libmongocrypt to ensure agreement.

* disable extra alignment in `linker-tests.sh`

To match libbson recommendeded defaults.

---------

Co-authored-by: Ezra Chung <88335979+eramongodb@users.noreply.github.com>