MONGOCRYPT-572 remove QE-V1 code and tests (#934)
* do not allow opting in to QEv1

`mongocrypt_setopt_fle2v2` can only be called privately. It is only used in now-defunct tests of QEv1.

* remove `_test_encrypt_fle2_delete_v1`

QEv2 does not append `deleteTokens`

* fix failing tests

Update payloads to V2 payloads. Remove `eccCollection`. The `eccCollection` is not used in QEv2.

* remove redundant tests

v1 tests were converted to v2. v1 is no longer needed.

* remove `TESTER_MONGOCRYPT_WITH_CRYPT_V1`

* remove `use_v2`

* remove unused test files

* remove remaining references to `eccCollection` in test data

* ignore `eccCollection`

The `eccCollection` is unused in QEv2

* remove deleteTokens

* remove QEv1 code in mongocrypt-marking.c

* remove `mongocrypt_setopt_fle2v2`

* test decrypting each payload type

* test FLE2InsertUpdatePayload with edges

Edges are sent for range payloads. Tested elsewhere. Redundant test is soon to be removed.

* remove redundant explicit decryption tests